backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
drf-spectacular==0.27.0
pytz==2023.3
setuptools>=65.0.0 